libshviotqt: Use int64_t for request ID

+++ libshviotqt/include/shv/iotqt/rpc/serialportsocket.h
@@ -23,8 +23,8 @@ public:
 	SerialFrameReader(CrcCheck crc);
 	~SerialFrameReader() override = default;
 
-	QList<int> addData(std::string_view data) override;
-	std::vector<std::pair<int, QString>> takeResponseErrors();
+	QList<int64_t> addData(std::string_view data) override;
+	std::vector<std::pair<int64_t, QString>> takeResponseErrors();
 	ReadState readState() const { return m_readState; }
 	void resetCommunication() override;
 private:
@@ -37,7 +37,7 @@ private:
 	std::string m_crcBuffer;
 	shv::chainpack::Crc32Shv3 m_crcDigest;
 	bool m_withCrcCheck = true;
-	std::vector<std::pair<int, QString>> m_responseErrors;
+	std::vector<std::pair<int64_t, QString>> m_responseErrors;
 };
 
 class LIBSHVIOTQT_EXPORT SerialFrameWriter : public FrameWriter
